chore: :truck: 회원가입 js파일들 디렉토리 이동
