+++ public/styles/error.css
@@ -1,4 +1,5 @@
 .error-image {
+  margin-top: 50px;
   width: 100%;
 }
