+++ views/pages/auth/signup-detail.pug
@@ -23,4 +23,4 @@ block content
       +input("생년월일", 'birth', "text" , "ex) 2000.01.01", "생년월일을 제대로 입력해주세요.", true, false, 10).birth
 
 block script
-  script(src='/js/signup-detail.js', type='module')
+  script(src='/js/signup/detail.js', type='module')
