+++ views/pages/auth/signup-phone.pug
@@ -26,4 +26,4 @@ block content
       p#signup-phone-reload-code.signup-phone-reload-code 인증번호 다시받기
 
 block script
-  script(src='/js/signup-phone.js', type='module')
+  script(src='/js/signup/phone.js', type='module')
