+++ views/pages/auth/signup-service-term.pug
@@ -29,4 +29,4 @@ block content
     +button("다음으로", "#bbbbbb")#service-term-next-button.service-term-next-button
 
 block script
-  script(src='/js/signup-service-term.js') 
+  script(src='/js/signup/service-term.js')
